+++ TelegramBotAPI-extensions-api/src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/extensions/api/GetUpdates.kt
@@ -2,14 +2,13 @@ package com.github.insanusmokrassar.TelegramBotAPI.extensions.api
 
 import com.github.insanusmokrassar.TelegramBotAPI.bot.RequestsExecutor
 import com.github.insanusmokrassar.TelegramBotAPI.requests.GetUpdates
-import com.github.insanusmokrassar.TelegramBotAPI.types.ALL_UPDATES_LIST
-import com.github.insanusmokrassar.TelegramBotAPI.types.UpdateIdentifier
+import com.github.insanusmokrassar.TelegramBotAPI.types.*
 import com.github.insanusmokrassar.TelegramBotAPI.types.update.abstracts.Update
 
 suspend fun RequestsExecutor.getUpdates(
     offset: UpdateIdentifier? = null,
     limit: Int? = null,
-    timeout: Int? = null,
+    timeout: Seconds? = null,
     allowed_updates: List<String>? = ALL_UPDATES_LIST
 ) = execute(
     GetUpdates(
@@ -20,7 +19,7 @@ suspend fun RequestsExecutor.getUpdates(
 suspend fun RequestsExecutor.getUpdates(
     lastUpdate: Update,
     limit: Int? = null,
-    timeout: Int? = null,
+    timeout: Seconds? = null,
     allowed_updates: List<String>? = ALL_UPDATES_LIST
 ) = getUpdates(
     lastUpdate.updateId + 1, limit, timeout, allowed_updates