+++ TelegramBotAPI/src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/bot/Ktor/base/AbstractRequestCallFactory.kt
@@ -1,8 +1,10 @@
 package com.github.insanusmokrassar.TelegramBotAPI.bot.Ktor.base
 
 import com.github.insanusmokrassar.TelegramBotAPI.bot.Ktor.KtorCallFactory
+import com.github.insanusmokrassar.TelegramBotAPI.requests.GetUpdates
 import com.github.insanusmokrassar.TelegramBotAPI.requests.abstracts.Request
 import io.ktor.client.HttpClient
+import io.ktor.client.features.timeout
 import io.ktor.client.request.*
 import io.ktor.client.statement.HttpStatement
 import io.ktor.http.ContentType
@@ -27,6 +29,16 @@ abstract class AbstractRequestCallFactory : KtorCallFactory {
                 )
                 method = HttpMethod.Post
                 accept(ContentType.Application.Json)
+
+                if (request is GetUpdates) {
+                    request.timeout ?.times(1000L) ?.let { customTimeoutMillis ->
+                        if (customTimeoutMillis > 0) {
+                            timeout {
+                                requestTimeoutMillis = customTimeoutMillis
+                            }
+                        }
+                    }
+                }
 
                 body = preparedBody
             },