+++ TelegramBotAPI/src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/requests/GetUpdates.kt
@@ -1,9 +1,7 @@
 package com.github.insanusmokrassar.TelegramBotAPI.requests
 
-import com.github.insanusmokrassar.TelegramBotAPI.bot.RequestsExecutor
 import com.github.insanusmokrassar.TelegramBotAPI.requests.abstracts.SimpleRequest
-import com.github.insanusmokrassar.TelegramBotAPI.types.ALL_UPDATES_LIST
-import com.github.insanusmokrassar.TelegramBotAPI.types.UpdateIdentifier
+import com.github.insanusmokrassar.TelegramBotAPI.types.*
 import com.github.insanusmokrassar.TelegramBotAPI.types.update.abstracts.Update
 import com.github.insanusmokrassar.TelegramBotAPI.types.update.abstracts.UpdateSerializerWithoutDeserialization
 import kotlinx.serialization.*
@@ -17,7 +15,7 @@ private val updatesListSerializer = ArrayListSerializer(
 data class GetUpdates(
     val offset: UpdateIdentifier? = null,// set `last update id + 1` to receive next part of updates
     val limit: Int? = null,
-    val timeout: Int? = null,
+    val timeout: Seconds? = null,
     val allowed_updates: List<String>? = ALL_UPDATES_LIST
 ): SimpleRequest<List<Update>> {
     override fun method(): String = "getUpdates"