+++ TelegramBotAPI/src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/update/abstracts/BaseSentMessageUpdate.kt
@@ -1,0 +1,3 @@
+package com.github.insanusmokrassar.TelegramBotAPI.types.update.abstracts
+
+interface BaseSentMessageUpdate : BaseMessageUpdate