+++ TelegramBotAPI/src/jvmMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/updateshandlers/KtorUpdatesCIOPoller.kt
@@ -4,13 +4,13 @@ import com.github.insanusmokrassar.TelegramBotAPI.bot.Ktor.KtorRequestsExecutor
 import com.github.insanusmokrassar.TelegramBotAPI.types.ALL_UPDATES_LIST
 import com.github.insanusmokrassar.TelegramBotAPI.types.update.abstracts.Update
 import com.github.insanusmokrassar.TelegramBotAPI.utils.TelegramAPIUrlsKeeper
-import com.github.insanusmokrassar.TelegramBotAPI.utils.extensions.UpdateReceiver
 import io.ktor.client.HttpClient
 import io.ktor.client.engine.HttpClientEngine
 import io.ktor.client.engine.cio.CIO
 import io.ktor.client.engine.cio.endpoint
 import io.ktor.util.KtorExperimentalAPI
 
+@Deprecated("Deprecated due to more simple way to get updates using TelegramBotAPI-extensions-api")
 @KtorExperimentalAPI
 fun KtorUpdatesPoller(
     telegramAPIUrlsKeeper: TelegramAPIUrlsKeeper,
@@ -44,6 +44,7 @@ fun KtorUpdatesPoller(
     )
 }
 
+@Deprecated("Deprecated due to more simple way to get updates using TelegramBotAPI-extensions-api")
 @KtorExperimentalAPI
 fun KtorUpdatesPoller(
     telegramAPIUrlsKeeper: TelegramAPIUrlsKeeper,